Module revolut_customer::amount

Revolut currency amount

This module holds the Amount type and the AmountParseError.

The maximum and minimum amount values can in any case be known by using max_value() and min_value() functions in the Amount type, or the MAX and MIN constants:

use std::u64;
use revolut_customer::Amount;

let max_value = Amount::max_value();
let min_value = Amount::min_value();

assert_eq!(max_value, Amount::from_repr(u64::max_value()));
assert_eq!(min_value, Amount::from_repr(u64::min_value()));

Structs

Amount
Revolut currency amount

Constants

MAX
Largest possible currency amount.

MIN
Smallest possible currency amount

_IMPL_DESERIALIZE_FOR_Amount

_IMPL_SERIALIZE_FOR_Amount