Struct revolut_customer::amount::Amount

pub struct Amount {
    value: u64,
}

Revolut currency amount

This data structure can be used the same way as any other number. An Amount can be added or subtracted to another Amount, and it can be divided and multiplied by an integer. All operations that are defined in the Amount scope and that are exact can be used directly as usual integer / float point operations.

No negative amounts can exist, since an Amount is unsigned, so the negation operator '-', then, has no use with an Amount.

Its internal representation is a 64 bit unsigned integer, that is displayed as a fixed point, number of factor 1/100. This means that an internal representation of 100 would be an external amount of 1. The internal representation shouldn't be used except when serializing and deserializing the data, since this type is sent in JSON as its internal u64.

The use is the following:

use revolut_customer::Amount;

let amount = Amount::from_repr(1_65); // 1.65
let ten = Amount::from_repr(10_00); // 10
let add_ten = amount + ten;
assert_eq!(add_ten, Amount::from_repr(11_65)); // 11.65

They can be divided and multiplied by any other unsigned integer:

let mut amount = Amount::from_repr(7_00); // 7
amount *= 10u32;
assert_eq!(amount, Amount::from_repr(70_00)); // 70

amount = amount / 30u16;
assert_eq!(amount, Amount::from_repr(2_33)); // 2.33

amount %= 1u8;
assert_eq!(amount, Amount::from_repr(0_33)); // 0.33

Amounts can easily be displayed using the Display trait as any other number:

let amount = Amount::from_repr(56_00);
assert_eq!(format!("{}", amount), "56");
assert_eq!(format!("{:.2}", amount), "56.00");
assert_eq!(format!("{:.5}", amount), "56.00000");
assert_eq!(format!("{:05.1}", amount), "056.0");

// And with rounding:
let amount = Amount::from_repr(0_56); // 0.56
assert_eq!(format!("{:.1}", amount), "0.6");

Fields
value: u64
Methods

impl Amount

pub fn from_repr(value: u64) -> Self

Creates a new amount from its internal representation.

pub fn get_repr(self) -> u64

Gets the internal representation of the amount.

pub const fn min_value() -> Self

Returns the smallest value that can be represented as a currency amount.

pub const fn max_value() -> Self

Returns the largest value that can be represented as a currency amount.