Struct revolut_customer::Client

pub struct Client {
    options: Options,
    client: Client,
    user_id: Option<Uuid>,
    access_token: Option<String>,
}

API client.

TODO: Client examples

Request API specification

The API is not publicly available, but reverse engineering what the Revolut iPhone/Android clients do, the following has been determined:

Headers:

The headers set by the application are the following:

X-Client-Version: client.version
X-Api-Version: 1
X-Device-Id: YOUR-DEVICE-ID
X-Device-Model: DeviceModel
User-Agent: AppUserAgent
Accept: application/json

The client sets the following defaults for the default (iPhone) configuration:

X-Client-Version: 5.29
X-Api-Version: 1
X-Device-Id: SOME-DEVICE-ID
X-Device-Model: iPhone8,1
User-Agent: Revolut/com.revolut.revolut (iPhone; iOS 11.1)
Accept: application/json

For the authenticated APIs, it uses simple authentication with the User ID as the user and the access token as the password, adding the header:

Authorization: Basic e3Jldm9sdXQtdXNlci1pZH06e0FjY2Vzc1Rva2VufQ==

The last part is the Base64 encoding of the {revolut-user-id}:{AccessToken} pair.

Fields
options: Options

Options for the client.

client: Client

HTTP client.

user_id: Option<Uuid>

Client ID.

access_token: Option<String>

Access token.

Methods

impl Client

Authorization client methods

pub fn sign_in<PH, PW>(&self, phone: PH, password: PW) -> Result<(), Error> where
PH: AsRef<str>,
PW: AsRef<str>,

Signs the user in.

Example:

use revolut_customer::{Client, ApiError};

let client = Client::default();
let response = client.sign_in("+1555555555", "9999");
assert_eq!(response.err().unwrap().downcast_ref::<ApiError>().unwrap(),
           &ApiError::Unauthorized);

Note that the response will be an unauthorized error, since the phone/password combination is not correct.

Request API specification

No authentication required.

GET https://api.revolut.com/signin

Body (JSON encoded):

{
    "phone": "+1555555555",
    "password": "9999"
}

The response status code will be in the 2XX range if the phone/password were correct, or in the 4XX range if they weren't or the API changed. The response will not have further information.

pub fn confirm_sign_in<P, C>(
&mut self,
phone: P,
code: C
) -> Result<(User, Wallet), Error> where
P: AsRef<str>,
C: AsRef<str>,

Confirms the user sign-in.

This will set the client with the user ID and the access token so that it can perform further requests that require authentication. That's the reason why the client needs to be mutable.

Example:

use revolut_customer::{Client, ApiError};

let mut client = Client::default();
let response = client.confirm_sign_in("+1555555555", "111-111");
assert!(response.is_err());

Note that the response will be a 400 error, since the phone/code combination is not correct.

Request API specification

No authentication required.

GET https://api.revolut.com/signin/confirm

Body (JSON encoded):

{
    "phone": "+1555555555",
    "code": "111-111"
}

The response status code will be in the 2XX range if the phone/code were correct, or in the 4XX range if they weren't or the API changed. If the response is correct, a JSON object containing the user, wallet and access token for the user si returned. The implementation only returns the user and wallet objects, and saves the access token and user ID to authenticate in future requests.

impl Client

User client methods.

They require the client to have loaded the authentication mechanisms.

pub fn current_user(&self) -> Result<(User, Wallet), Error>

Gets user information.

Make sure the client has the authentication information.

pub fn current_user_wallet(&self) -> Result<Wallet, Error>

Gets user's wallet information.

Make sure the client has the authentication information.

pub fn current_user_cards(&self) -> Result<Vec<Card>, Error>

Gets user's cards information.

Make sure the client has the authentication information.

pub fn change_current_user_address(
&self,
address: &Address
) -> Result<(), Error>

Changes the address of the current user.

This method will set the address of the user to the given one. Note: Make sure the client has the authentication information.

Example:

use std::env;
use revolut_customer::{Client, private::Address};

let mut client = Client::default();

let user_id = env::var("TEST_USER_ID")
    .expect("TEST_USER_ID environment variable not set");
let access_token = env::var("TEST_ACCESS_TOKEN")
    .expect("TEST_ACCESS_TOKEN environment variable not set");

client
   .set_auth(user_id, access_token)
   .expect("invalid user ID");

let new_address = Address::new(
    "NewCity",
    "FR",
    "39325",
    "NewRegion",
    "Street 1, 6",
    None);
client.change_current_user_address(&new_address).unwrap();

let (new_user, _wallet) = client.current_user().unwrap();
assert_eq!(new_user.address(), &new_address);

Note that the response will be a 400 error, since the phone/code combination is not correct.

Request API specification

No authentication required.

GET https://api.revolut.com/signin/confirm

Body (JSON encoded):

{
    "phone": "+1555555555",
    "code": "111-111"
}

The response status code will be in the 2XX range if the phone/code were correct, or in the 4XX range if they weren't or the API changed. If the response is correct, a JSON object containing the user, wallet and access token for the user si returned. The implementation only returns the user and wallet objects, and saves the access token and user ID to authenticate in future requests.

The definitions for these objects is shown in the methods that specifically return each of the types.

impl Client

Public client methods.

impl Client

pub fn with_options(options: Options) -> Self

Creates a new client with the given options.

pub fn set_options(&mut self, options: Options)

Changes the options of the client.

pub fn set_auth<I, T>(
&mut self,
user_id: I,
access_token: T
) -> Result<(), Error> where
I: AsRef<str>,
T: Into<String>,

Sets the user authentication information for the client.

pub fn user_id(&self) -> Option<Uuid>

Gets the logged in access token.

pub fn access_token(&self) -> Option<&String>

Gets the logged in access token.

pub fn unset_auth(&mut self)

Removes the user authentication information.

This is effectively logging the user out.

fn set_headers(&self, request_builder: RequestBuilder) -> RequestBuilder

Sets the headers with the provided documentation.